Improve support for cache clearing
Allows for full cache clearing (including child sitemaps), or individual child sitemaps.

Refactored code to the Sitemapamic support class for centralisation, and updated Listener to clear the entire Sitemapamic Cache when changes are made.
