+++ src/Listeners/ClearSitemapamicCache.php
@@ -4,14 +4,15 @@ namespace MityDigital\Sitemapamic\Listeners;
 
 use Illuminate\Contracts\Queue\ShouldQueue;
 use Illuminate\Support\Facades\Cache;
+use MityDigital\Sitemapamic\Facades\Sitemapamic;
 
 class ClearSitemapamicCache implements ShouldQueue
 {
     /**
-     * Simply clear the Sitemapamic cache
+     * Simply clear the entire Sitemapamic cache
      */
     public function handle()
     {
-        Cache::forget(config('sitemapamic.cache'));
+        Sitemapamic::clearCache();
     }
 }
